backend/requirements.txt
Django==4.2.19
djangorestframework==3.15.2
django-filter==23.2
djoser==2.3.1
drf-extra-fields==3.7.0
gunicorn==23.0.0
pillow==11.1.0
python-dotenv==1.0.0